# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"auth": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					},
					"SAPCode": {
						"type": "string"
					},
					"SAPUser": {
						"type": "string"
					},
					"requiredPassUpdate": {
						"type": "boolean"
					},
					"datePassUpdate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"username",
					"password",
					"email",
					"role",
					"isActive",
					"auth",
					"created",
					"updatedDate",
					"operation",
					"area",
					"line",
					"SAPCode",
					"SAPUser",
					"requiredPassUpdate",
					"datePassUpdate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Operation": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"bus": {
						"$ref": "#/components/schemas/Bus"
					},
					"code": {
						"type": "string"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"bus",
					"code",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Area": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate",
					"operation",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Line": {
				"properties": {
					"id": {
						"type": "string"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"area",
					"name",
					"SAPCode",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate",
					"operation",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Card": {
				"properties": {
					"id": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"process": {
						"$ref": "#/components/schemas/Process"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"groupCode",
					"name",
					"operation",
					"process",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Priority": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"operation",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Component": {
				"properties": {
					"id": {
						"type": "string"
					},
					"machine": {
						"$ref": "#/components/schemas/LineMachine"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"machine",
					"name",
					"SAPCode",
					"operation",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Breakdown": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"operation",
					"userUpdate",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TypeFail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"shortName": {
						"type": "string"
					},
					"activity": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"isActive",
					"shortName",
					"activity",
					"operation",
					"userUpdate",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Notice": {
				"properties": {
					"id": {
						"type": "string"
					},
					"numNotice": {
						"type": "number",
						"format": "double"
					},
					"otCode": {
						"type": "string"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"departmentId": {
						"$ref": "#/components/schemas/Area"
					},
					"lineId": {
						"$ref": "#/components/schemas/Line"
					},
					"cardTypeId": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priorityId": {
						"$ref": "#/components/schemas/Priority"
					},
					"componentsId": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdownId": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureTypeId": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"affectsId": {
						"$ref": "#/components/schemas/Affect"
					},
					"affectsFile": {
						"type": "string"
					},
					"cardDescription": {
						"type": "string"
					},
					"processId": {
						"$ref": "#/components/schemas/Process"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"sapId": {
						"type": "string"
					},
					"equipmentId": {
						"$ref": "#/components/schemas/LineMachine"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"responsableId": {
						"$ref": "#/components/schemas/Responsable"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					},
					"urlPhoto": {
						"type": "string"
					},
					"userUpdate": {
						"$ref": "#/components/schemas/User"
					},
					"failureTimeStartDate": {
						"type": "string"
					},
					"failureTimeStartTime": {
						"type": "string"
					},
					"startHour": {
						"type": "string"
					},
					"endHour": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"otCode",
					"didCard",
					"failureTime",
					"departmentId",
					"lineId",
					"cardTypeId",
					"cardTitle",
					"priorityId",
					"componentsId",
					"breakdownId",
					"failureTypeId",
					"affectsId",
					"affectsFile",
					"cardDescription",
					"processId",
					"user",
					"sapId",
					"equipmentId",
					"operation",
					"objectId",
					"responsableId",
					"causeId",
					"symptomId",
					"textCause",
					"textSymptom",
					"urlPhoto",
					"userUpdate",
					"failureTimeStartDate",
					"failureTimeStartTime",
					"startHour",
					"endHour",
					"isActive",
					"created",
					"updatedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INoticePayload": {
				"properties": {
					"otCode": {
						"type": "string"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					},
					"cardType": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/Priority"
					},
					"components": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdown": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureType": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"affects": {
						"$ref": "#/components/schemas/Affect"
					},
					"affectsFile": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"cardDescription": {
						"type": "string"
					},
					"process": {
						"$ref": "#/components/schemas/Process"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"sapId": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					},
					"failureTimeStartDate": {
						"type": "string"
					},
					"failureTimeStartTime": {
						"type": "string"
					},
					"startHour": {
						"type": "string"
					},
					"endHour": {
						"type": "string"
					}
				},
				"required": [
					"otCode",
					"didCard",
					"failureTime",
					"department",
					"line",
					"cardType",
					"cardTitle",
					"priority",
					"components",
					"breakdown",
					"failureType",
					"affects",
					"affectsFile",
					"cardDescription",
					"process",
					"user",
					"sapId",
					"operation",
					"objectId",
					"causeId",
					"symptomId",
					"textCause",
					"textSymptom",
					"failureTimeStartDate",
					"failureTimeStartTime",
					"startHour",
					"endHour"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INoticenPayloadNewFormat": {
				"properties": {
					"processId": {
						"$ref": "#/components/schemas/Process"
					},
					"didCard": {
						"type": "string"
					},
					"failureTime": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"lineId": {
						"$ref": "#/components/schemas/Line"
					},
					"equipment": {
						"$ref": "#/components/schemas/Equipment"
					},
					"cardTypeId": {
						"$ref": "#/components/schemas/Card"
					},
					"cardTitle": {
						"type": "string"
					},
					"priorityId": {
						"$ref": "#/components/schemas/Priority"
					},
					"componentsId": {
						"$ref": "#/components/schemas/Component"
					},
					"breakdownId": {
						"$ref": "#/components/schemas/Breakdown"
					},
					"failureTypeId": {
						"$ref": "#/components/schemas/TypeFail"
					},
					"cardDescription": {
						"type": "string"
					},
					"affectsId": {
						"$ref": "#/components/schemas/Affect"
					},
					"otCode": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"affectsFile": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"isActive": {
						"type": "boolean",
						"enum": [
							"true"
						],
						"nullable": false
					},
					"sapId": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"objectId": {
						"$ref": "#/components/schemas/ObjectParts"
					},
					"causeId": {
						"$ref": "#/components/schemas/Cause"
					},
					"symptomId": {
						"$ref": "#/components/schemas/Symptom"
					},
					"textCause": {
						"type": "string"
					},
					"textSymptom": {
						"type": "string"
					}
				},
				"required": [
					"processId",
					"didCard",
					"failureTime",
					"department",
					"lineId",
					"equipment",
					"cardTypeId",
					"cardTitle",
					"priorityId",
					"componentsId",
					"breakdownId",
					"failureTypeId",
					"cardDescription",
					"affectsId",
					"otCode",
					"id",
					"affectsFile",
					"userId",
					"isActive",
					"sapId",
					"operation",
					"textCause",
					"textSymptom"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INoticeThirdParties": {
				"properties": {
					"noticeType": {
						"type": "string"
					},
					"codification": {
						"type": "string"
					},
					"priority": {
						"type": "string"
					},
					"repercussion": {
						"type": "string"
					},
					"plannerGroup": {
						"type": "string"
					},
					"equipment": {
						"type": "string"
					},
					"technicalLocation": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"objectPartText": {
						"type": "string"
					},
					"symptomText": {
						"type": "string"
					},
					"equipmentSet": {
						"type": "string"
					},
					"operation": {
						"type": "string"
					}
				},
				"required": [
					"noticeType",
					"codification",
					"priority",
					"repercussion",
					"plannerGroup",
					"equipment",
					"technicalLocation",
					"title",
					"description",
					"objectPartText",
					"symptomText",
					"equipmentSet",
					"operation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fixmaint",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Richard HC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users/{id}": {
			"put": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/notices": {
			"get": {
				"operationId": "GetNotices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Notice"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "top",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dateFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateEnd",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sapForm",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isWeb",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "timeFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeEnd",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "operationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "totalRows",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "timezone",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNoticeNewFormat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Notice"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticenPayloadNewFormat"
							}
						}
					}
				}
			}
		},
		"/notices/old_notice": {
			"post": {
				"operationId": "CreateNotice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Notice"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticePayload"
							}
						}
					}
				}
			}
		},
		"/notices/external-notice-consumers": {
			"post": {
				"operationId": "CreateNoticeThirdParties",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										},
										"resData": {},
										"reqData": {
											"$ref": "#/components/schemas/INoticeThirdParties"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"msg",
										"resData",
										"reqData",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticeThirdParties"
							}
						}
					}
				}
			}
		},
		"/notices/notice-machine": {
			"get": {
				"operationId": "GetNoticesByMachineId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Notice"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notices/{id}": {
			"put": {
				"operationId": "UpdateNotice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INoticenPayloadNewFormat"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetNotice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Notice"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Notice"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}